refactor: Move date picker to a new component
- Add dayjs dependency
