+++ src/components/leaderboard.tsx
@@ -28,13 +28,9 @@ import Github from "../../public/icons/github.svg";
 import Telegram from "../../public/icons/telegram.svg";
 import { Badge } from "./ui/badge";
 import { Skeleton } from "./ui/skeleton";
-import { CalendarIcon } from "lucide-react";
-import { Button } from "@/components/ui/button";
-import { Calendar } from "@/components/ui/calendar";
-import { format } from "date-fns"
-import { Popover, PopoverContent, PopoverTrigger, } from "@/components/ui/popover";
 import { DateRange } from "react-day-picker";
-
+import DatePickerWithPresets from "@/components/date-picker";
+import dayjs from 'dayjs';
 
 interface LeaderboardProps {
   data: LeaderboardEntry[] | null;
@@ -55,10 +51,6 @@ interface LeaderboardProps {
   }[];
   onTokenSelect?: (tokenId: string) => void;
   slug: string;
-}
-
-interface DatePickerProps {
-  onDateChange: (date: DateRange | null) => void;
 }
 
 function LeaderboardHeader({
@@ -157,145 +149,6 @@ const EmptyState = ({ metadata }: Pick<LeaderboardProps, "metadata">) => {
   );
 };
 
-const DatePickerWithPresets = ({onDateChange}: DatePickerProps) => {
-  const today = new Date();
-  const t = useTranslations( "overview.leaderboard" );
-  const [date, setDate] = useState<DateRange | undefined>( {
-    from: new Date(new Date().setDate(today.getDate() - 30)),
-    to: new Date(),
-  } );
-  const [preset, setPreset] = useState<string>( "0" );
-  const [startMonth, setStartMonth] = useState( new Date(new Date().setDate(today.getDate() - 30)) );
-  const [endMonth, setEndMonth] = useState( new Date() );
-  const presets = {
-    "0": "Last month",
-    "1": "This month",
-    "2": "Three months",
-    "3": "All time",
-    "4": "Custom",
-  };
-
-  const daysAgo = (n: number) => {
-    const d = new Date();
-    d.setDate( d.getDate() - Math.abs( n ) );
-    return d;
-  };
-
-  const onDayClick = (day: DateRange) => {
-    setDate( day );
-    setPreset( "4" );
-  };
-
-  useEffect( () => {
-    if (date?.to && date.from && onDateChange) {
-      onDateChange( date );
-    }
-  }, [date] );
-
-  return (
-    <Popover>
-      <PopoverTrigger asChild>
-        <Button
-          variant={"outline"}
-          className={cn(
-            "w-[240px] justify-start text-left font-normal",
-            !date && "text-muted-foreground"
-          )}
-        >
-          <CalendarIcon/>
-          {date?.from ? (
-            date.to ? (
-              <>
-                {format( date.from, "LLL dd, y" )} -{" "}
-                {format( date.to, "LLL dd, y" )}
-              </>
-            ) : (
-              format( date.from, "LLL dd, y" )
-            )
-          ) : (
-            <span>{t( "datePicker.pickDate" )}</span>
-          )}
-        </Button>
-      </PopoverTrigger>
-      <PopoverContent
-        align="start"
-        className="flex w-auto flex-col space-y-2 p-2"
-      >
-        <Select
-          value={preset}
-          onValueChange={(value) => {
-            if (value === '0') {
-              const newFrom = daysAgo( 30 );
-              const newDate: DateRange = {
-                from: newFrom,
-                to: today,
-              };
-              setDate( newDate );
-              setStartMonth( newFrom );
-              setEndMonth( today );
-              setPreset( "0" );
-            } else if (value === '1') {
-              const thisMonth = new Date( today.getFullYear(), today.getMonth(), 1 );
-              const newDate: DateRange = {
-                from: thisMonth,
-                to: today,
-              };
-              setDate( newDate );
-              setStartMonth( thisMonth );
-              setPreset( "1" );
-            } else if (value === '2') {
-              const newFrom = daysAgo( 90 );
-              const newDate: DateRange = {
-                from: newFrom,
-                to: new Date(),
-              };
-              setDate( newDate );
-              setStartMonth( newFrom )
-              setEndMonth( today );
-              setPreset( "2" );
-            } else if (value === '3') {
-              const newDate: DateRange = {
-                from: new Date( 2022, 3, 2 ),
-                to: new Date(),
-              };
-              setDate( newDate );
-              setStartMonth( new Date( 2022, 3, 2 ) )
-              setEndMonth( today );
-              setPreset( "3" );
-            }
-          }
-          }
-        >
-          <SelectTrigger>
-            <SelectValue placeholder={presets[preset]}/>
-          </SelectTrigger>
-          <SelectContent position="popper">
-            <SelectItem value="4">{t( "datePicker.custom" )}</SelectItem>
-            <SelectItem value="0">{t( "datePicker.lastMonth" )}</SelectItem>
-            <SelectItem value="1">{t( "datePicker.thisMonth" )}</SelectItem>
-            <SelectItem value="2">{t( "datePicker.threeMonths" )}</SelectItem>
-            <SelectItem value="3">{t( "datePicker.allTime" )}</SelectItem>
-          </SelectContent>
-        </Select>
-        <div className="rounded-md border">
-          <Calendar
-            initialFocus={true}
-            mode="range"
-            month={startMonth}
-            endMonth={endMonth}
-            selected={date}
-            onSelect={onDayClick}
-            onMonthChange={setStartMonth}
-            numberOfMonths={2}
-            required
-            disabled={{after: new Date(), before: new Date( 2022, 3, 2 )}}
-          />
-        </div>
-      </PopoverContent>
-    </Popover>
-  )
-};
-
 export default function Leaderboard({
   data,
   metadata,
@@ -331,7 +184,9 @@ export default function Leaderboard({
   const handleTokenSelect = async (tokenId: string) => {
     setIsLoading(true);
     try {
-      const newData = await generateLeaderboard(slug, tokenId, date?.from?.getTime().toString(), date?.to?.getTime().toString());
+      const startDate = dayjs(date?.from).unix().toString();
+      const endDate = dayjs(date?.to).unix().toString();
+      const newData = await generateLeaderboard(slug, tokenId, startDate, endDate);
       setLocalData(newData);
       setSelectedTokenId(tokenId);
     } finally {
